% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_info}
\alias{cache_info}
\title{Display information about the openesm cache}
\usage{
cache_info()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Shows the location and total size of the local file cache.
}
\examples{
\dontshow{
.tmp_cache <- tempfile("openesm_ex_")
.old <- options(openesm.cache_dir = .tmp_cache)
}
\donttest{
# view cache information
cache_info()
}
\dontshow{
options(.old)
unlink(.tmp_cache, recursive = TRUE)
}
}
