% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{cite.openesm_dataset}
\alias{cite.openesm_dataset}
\title{Cite method for openesm_dataset}
\usage{
\method{cite}{openesm_dataset}(x, format = "bibtex", ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{format}{Character, format for citation (currently only enables "bibtex").}

\item{...}{Additional arguments (not used).}
}
\value{
A character string with the citation information.
}
\description{
Cite method for openesm_dataset
}
