% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cite}
\alias{cite}
\title{Citation information for openESM datasets}
\usage{
cite(x, ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{...}{Additional arguments passed to methods}
}
\value{
Citation information for the dataset(s)
}
\description{
Citation information for openESM datasets
}
\examples{
\donttest{
dataset <- structure(
  list(
    dataset_id = "0001",
    metadata = list(
      reference_a = "@article{fried2022, author = {Fried}, year = {2022}}"
    )
  ),
  class = "openesm_dataset"
)
cite(dataset)
}
}
