% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{download_from_zenodo}
\alias{download_from_zenodo}
\title{Download dataset from Zenodo.}
\usage{
download_from_zenodo(
  zenodo_doi,
  dataset_id,
  author_name,
  version,
  sandbox = FALSE,
  dest_path = NULL
)
}
\arguments{
\item{zenodo_doi}{Character string with the Zenodo concept DOI}

\item{dataset_id}{Character string with dataset identifier}

\item{author_name}{Character string with author name}

\item{version}{Character string specifying a specific version tag (e.g., "v1.0.0").}

\item{sandbox}{Logical, whether to use Zenodo sandbox. Defaults to FALSE. Mainly used for  internal testing}

\item{dest_path}{Character string with destination path}
}
\value{
Character string with path to downloaded file.
}
\description{
Download dataset from Zenodo.
}
