% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{is_openesm_dataset}
\alias{is_openesm_dataset}
\title{Check if object is an openesm_dataset}
\usage{
is_openesm_dataset(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is an openesm_dataset
}
