% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{notes.openesm_dataset}
\alias{notes.openesm_dataset}
\title{Notes method for openesm_dataset}
\usage{
notes.openesm_dataset(x, ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{...}{Additional arguments (not used).}
}
\value{
A character vector with the notes for the dataset, returned invisibly.
}
\description{
Notes method for openesm_dataset
}
