% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{validate_openesm_dataset}
\alias{validate_openesm_dataset}
\title{Validate openesm_dataset object}
\usage{
validate_openesm_dataset(x)
}
\arguments{
\item{x}{Object to validate}
}
\description{
Validate openesm_dataset object
}
\keyword{internal}
