% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cite}
\alias{cite}
\title{Citation information for OpenESM datasets}
\usage{
cite(x, ...)
}
\arguments{
\item{x}{An object to get citation information for}

\item{...}{Additional arguments passed to methods}
}
\description{
Citation information for OpenESM datasets
}
