% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the openesm cache}
\usage{
clear_cache(force = FALSE)
}
\arguments{
\item{force}{Logical, if TRUE, will not ask for confirmation before deleting.
Defaults to FALSE.}
}
\description{
Removes all cached openesm data from your local machine.
}
