% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a dataset from OpenESM}
\usage{
get_dataset(
  dataset_id,
  version = NULL,
  cache = TRUE,
  path = NULL,
  force_download = FALSE
)
}
\arguments{
\item{dataset_id}{Character string or vector of dataset IDs}

\item{version}{Character string specifying the version (default is "latest")}

\item{cache}{Logical, if TRUE uses cached version if available (default is TRUE)}

\item{path}{Character string specifying the path to save the dataset (default is NULL)}

\item{force_download}{Logical, if TRUE forces re-download even if cached version exists (
default is FALSE)}
}
\value{
A data frame with the dataset, or a list of data frames for multiple
datasets
}
\description{
Get a dataset from OpenESM
}
\examples{
\dontrun{
# Get a single dataset
dataset <- get_dataset("example_dataset_id")

# Get multiple datasets
datasets <- get_dataset(c("dataset1", "dataset2"))
}
}
