% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List Available Datasets
#' Retrieves a list of available datasets from the OpenESM metadata repository.
It checks for a cached version of the dataset index and downloads a fresh copy if the
cache is older than the specified number of hours.}
\usage{
list_datasets(cache_hours = 24)
}
\arguments{
\item{cache_hours}{Numeric, number of hours to consider the cache valid (default is 24).}
}
\value{
A tibble with one row per dataset, including a nested tibble with
detailed feature information.
}
\description{
List Available Datasets
#' Retrieves a list of available datasets from the OpenESM metadata repository.
It checks for a cached version of the dataset index and downloads a fresh copy if the
cache is older than the specified number of hours.
}
