% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{notes}
\alias{notes}
\title{Additional notes for OpenESM datasets}
\usage{
notes(x, ...)
}
\arguments{
\item{x}{An object to get notes for}

\item{...}{Additional arguments passed to methods}
}
\description{
Additional notes for OpenESM datasets
}
