% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.openesm_dataset}
\alias{print.openesm_dataset}
\title{S3 Methods for openesm objects
Print method for openesm_dataset}
\usage{
\method{print}{openesm_dataset}(x, ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Invisibly returns the original object.
}
\description{
S3 Methods for openesm objects
Print method for openesm_dataset
}
